\name{SUE-package}
\alias{SUE-package}
\alias{SUE}
\docType{package}
\title{
Subsampling method
}
\description{
This is a package for the subsampling method of robust estimation of linear regression models
}
\details{
\tabular{ll}{
Package: \tab SUE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-28\cr
License: \tab GPL-2\cr
}

}
\author{
Jim Yi

Maintainer: Jim Yi <yijinpeng@live.cn>
}

\keyword{ package }

