\name{SUE.fitted.values}
\alias{SUE.fitted.values}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fitted Values of SUE.lm Fits
}
\description{
Computes fitted values of a fit \code{\link{SUE.lm}} model.
}
\usage{
SUE.fitted.values(fit)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fit}{
 An object of class \code{SUE.lm} , typically the result of a call to \code{\link{SUE.lm}}
}
}
\author{
Jim Yi
}
\examples{
##See examples of the function SUE.lm
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
