\name{SUE.residuals}
\alias{SUE.residuals}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Residuals of SUE.lm Fits
}
\description{
Computes residuals of fit \code{\link{SUE.lm}} model.
}
\usage{
SUE.residuals(fit)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fit}{
 An object of class \code{SUE.lm} , typically the result of a call to \code{\link{SUE.lm}}
}
}
\author{
Jim Yi
}
\examples{
##See examples of the function SUE.lm
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line